move LogCollector to datadog.trace.api.telemetry
