+++ internal-api/src/main/java/datadog/trace/api/telemetry/LogCollector.java
@@ -1,4 +1,4 @@
-package datadog.trace.api;
+package datadog.trace.api.telemetry;
 
 import java.util.ArrayList;
 import java.util.Collection;
